数据结构变动

+++ src/main/java/com/lvmaoya/blog/domain/entity/RolePermission.java
@@ -12,9 +12,9 @@ public class RolePermission implements Serializable {
     /**
     * 角色ID
     */
-    private Long roleId;
+    private Integer roleId;
     /**
     * 权限ID
     */
-    private Long permissionId;
+    private Integer permissionId;
 }
